Move files to better places.
