+++ example-trace-messages/cons/__init__.py
@@ -1,3 +1,4 @@
 from .cons import mkenv
+from .prompt_to_md import prompt_to_md
 
-__all__ = ["mkenv"]
+__all__ = ["mkenv", "prompt_to_md"]
