+++ example-trace-messages/cons/nodes/stdout.py
