--- example-trace-messages/nodes/__init__.py
@@ -1,3 +1,0 @@
-from .prompt_to_md import prompt_to_md
-
-__all__ = ["prompt_to_md"]
