+++ example-trace-messages/start.py
@@ -1,5 +1,4 @@
-from nodes import prompt_to_md
-from cons import mkenv
+from cons import mkenv, prompt_to_md
 import os
 import json
